wip on add sale page

+++ app/Http/Controllers/Purchase/PurchaseController.php
@@ -10,6 +10,7 @@ use Illuminate\Contracts\View\View;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -28,6 +29,8 @@ class PurchaseController extends Controller
      */
     public function index(): View
     {
+        Gate::authorize('app.purchase.index');
+
         return view('pages.purchase.index', $this->services->accessPermissions());
     }
 
@@ -38,6 +41,8 @@ class PurchaseController extends Controller
      */
     public function getPurchaseList(): LengthAwarePaginator
     {
+        Gate::authorize('app.purchase.index');
+
         return $this->services->getPurchaseList();
     }
 
@@ -46,6 +51,8 @@ class PurchaseController extends Controller
      */
     public function create(): View
     {
+        Gate::authorize('app.purchase.create');
+
         return view('pages.purchase.create');
     }
 
@@ -56,6 +63,8 @@ class PurchaseController extends Controller
      */
     public function getProducts(): mixed
     {
+        Gate::authorize('app.purchase.create');
+
         return $this->services->getProducts();
     }
 
@@ -64,6 +73,8 @@ class PurchaseController extends Controller
      */
     public function getSuppliers(): Collection|array
     {
+        Gate::authorize('app.purchase.create');
+
         return $this->services->getSuppliers();
     }
 
@@ -72,6 +83,8 @@ class PurchaseController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        Gate::authorize('app.purchase.create');
+
         return $this->services
             ->validateStore($request)
             ->store($request);
@@ -82,6 +95,8 @@ class PurchaseController extends Controller
      */
     public function show(string $id): Purchase|array
     {
+        Gate::authorize('app.purchase.show');
+
         return $this->services->showDetails($id);
     }
 
@@ -91,6 +106,8 @@ class PurchaseController extends Controller
      */
     public function printPurchase(string $id): View
     {
+        Gate::authorize('app.purchase.show');
+
         return view('pages.purchase.print-purchase',['purchase' => (object) $this->services->showDetails($id)]);
     }
 
@@ -99,6 +116,8 @@ class PurchaseController extends Controller
      */
     public function edit(string $id): View
     {
+        Gate::authorize('app.purchase.edit');
+
         return view('pages.purchase.edit', ['id' => $id]);
     }
 
@@ -107,6 +126,8 @@ class PurchaseController extends Controller
      */
     public function update(Request $request, string $id): JsonResponse
     {
+        Gate::authorize('app.purchase.edit');
+
         return $this->services
             ->validateUpdate($request, $id)
             ->update($request, $id);
@@ -117,6 +138,8 @@ class PurchaseController extends Controller
      */
     public function destroy(string $id): JsonResponse
     {
+        Gate::authorize('app.purchase.delete');
+
         return $this->services->delete($id);
     }
 }
